Remove MLX format conversion for Wan2.2 T5 model
